requirements.txt
openai==1.3.0
pandas==2.1.0
python-dotenv==1.0.0
httpx==0.24.1
yfinance==0.2.30
alpha_vantage==3.0.0
